fixed logo on signup and login, moved logo and profile to assets
